Digest refactored
