+++ src/main/kotlin/ru/zelginni/tinycerberusbot/bot/BotController.kt
@@ -7,11 +7,12 @@ import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
+import ru.zelginni.tinycerberusbot.digest.DigestPublisher
 
 @RestController
 @RequestMapping("/admin/bot")
 class BotController(
-    private val tinyCerberusBot: TinyCerberusBot
+    private val digestPublisher: DigestPublisher,
 ) {
     private val logger = LoggerFactory.getLogger(BotController::class.java)
 
@@ -23,7 +24,7 @@ class BotController(
     )
     fun dailyDigest(): ResponseEntity<String> {
         logger.info("Launch digest requested")
-        tinyCerberusBot.dailyDigest()
+        digestPublisher.publishDailyDigest()
         return ResponseEntity.ok("Daily digest performed")
     }
 }
