+++ src/main/kotlin/ru/zelginni/tinycerberusbot/bot/TinyCerberusBot.kt
@@ -1,6 +1,5 @@
 package ru.zelginni.tinycerberusbot.bot
 
-import org.springframework.scheduling.annotation.Scheduled
 import org.springframework.stereotype.Component
 import org.telegram.telegrambots.meta.api.objects.Update
 import org.telegram.telegrambots.meta.api.objects.User
@@ -8,8 +7,6 @@ import org.telegram.telegrambots.meta.api.objects.message.Message
 import ru.zelginni.tinycerberusbot.bayan.Bayan
 import ru.zelginni.tinycerberusbot.bayan.BayanService
 import ru.zelginni.tinycerberusbot.chat.ChatService
-import ru.zelginni.tinycerberusbot.digest.DigestService
-import ru.zelginni.tinycerberusbot.chat.ChatViewDto
 import ru.zelginni.tinycerberusbot.rules.RulesService
 import ru.zelginni.tinycerberusbot.telegram.IncomingChatMessage
 import ru.zelginni.tinycerberusbot.telegram.TelegramUpdateHandler
@@ -20,10 +17,8 @@ import java.time.Instant
 class TinyCerberusBot(
     private val telegramCommandSender: TelegramCommandSender,
     private val telegramUpdateHandler: TelegramUpdateHandler,
-    private val botProperties: TelegramBotProperties,
     private val bayanService: BayanService,
     private val chatService: ChatService,
-    private val digestService: DigestService,
     private val rulesService: RulesService
 ) {
 
@@ -58,41 +53,6 @@ class TinyCerberusBot(
 
     private fun processCommand(update: Update) {
         telegramUpdateHandler.handleMessage(update.message.toIncomingChatMessage())
-    }
-
-    @Scheduled(cron = "\${bot.digest.cron}")
-    fun dailyDigest() {
-        val allChats: List<ChatViewDto> = chatService.getAllChats()
-        allChats.filter {
-                chat -> chat.enabled == true
-                && chat.digestEnabled == true
-                && chat.id != null
-                && chat.telegramId != null
-        }.forEach { chat -> sendDigest(chat)}
-        cleanDigestPins()
-    }
-
-    private fun sendDigest(chat: ChatViewDto) {
-        val chatId = chat.telegramId ?: return
-        val digest = digestService.compileDigest(chat.id ?: -1) ?: return
-        if (digest.isBlank()) {
-            telegramCommandSender.sendMessage(chatId.toLong(), "За прошедшие сутки в дайджест ничего не добавили.")
-            return
-        }
-        val message = telegramCommandSender.sendSilentMessage(chatId, digest) ?: return
-        val messageId = message.messageId ?: return
-        telegramCommandSender.pinMessage(message.chatId.toString(), messageId, true)
-        digestService.addPinnedDigest(chatId, messageId)
-        digestService.deleteDigest(chatId)
-    }
-
-    private fun cleanDigestPins() {
-        digestService.fetchOutdatedDigests()
-            ?.filter { it.chat?.telegramId != null && it.pinnedMessageId != null }
-            ?.forEach {
-                telegramCommandSender.unpinMessage(it.chat?.telegramId ?: "", it.pinnedMessageId ?: return@forEach)
-        }
-        digestService.deletePinnedDigests()
     }
 
     private fun sendSimpleReplyText(update: Update, messageText: String) {
@@ -177,5 +137,3 @@ class TinyCerberusBot(
     private fun User.writableName(): String =
         userName ?: "$firstName ${lastName ?: ""}".trim()
 }
-
-private const val ALT_COMMAND_START = "!"
